pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "urbansolarcarver"
version = "0.9.0"
description = "GPU-accelerated generation of maximum buildable volumes that respect solar access, daylight, and thermal comfort constraints in urban environments."
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.10"
keywords = ["solar-envelope", "urban-design", "daylight", "GPU", "voxel", "computational-design"]
authors = [
  { name = "Aristotelis Vartholomaios", email = "avartholomaios@uth.gr" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Education",
  "Topic :: Scientific/Engineering",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
]
dependencies = [
  "typer>=0.16.0",
  "pydantic>=2.0",
  "PyYAML>=6.0.2",
  "numpy>=1.24.3",
  "matplotlib>=3.10",
  "scipy>=1.10,<1.16",
  "numba>=0.58",
  "scikit-image>=0.24,<0.26",
  "Shapely>=2.1.1",
  "trimesh>=4.6.12",
  "rtree>=1.0",
  "ladybug-core>=0.44.19",
  "ladybug-radiance>=0.2.11",
  "warp-lang>=1.7.2",
  "torch>=2.0",  # CPU wheel is fine; setup_env.py installs the correct CUDA variant first
]

[project.urls]
Homepage = "https://github.com/avarth/UrbanSolarCarver"
Documentation = "https://avarth.github.io/UrbanSolarCarver/"
Repository = "https://github.com/avarth/UrbanSolarCarver"
Issues = "https://github.com/avarth/UrbanSolarCarver/issues"

[project.optional-dependencies]
cuda = [
  "torch>=2.0",
]
dev = ["pytest>=7.0"]
docs = [
  "mkdocs-material>=9.5",
  "mkdocstrings[python]>=0.24",
  "mkdocs-literate-nav>=0.6",
]

[project.scripts]
urbansolarcarver = "urbansolarcarver.carver_cli:app"
usc = "urbansolarcarver.carver_cli:app"

[tool.setuptools.packages.find]
where = ["src"]
include = ["urbansolarcarver*"]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true
